Update patch set 4
Patch Set 4:

Jake: You're right that the commit was for cm-10.1, but for whatever reason, CM didn't start fresh for that repo this time around. They piled all of the AOSP 4.4 changes on top of their existing commits.

If you look at the current copy of the code on the cm-11.0 branch, the changes that commit made are still present: https://github.com/CyanogenMod/android_external_libnfc-nci/blob/cm-11.0/Android.mk

Patch-set: 4