Update patch set 6
Patch Set 6:

Yea, as soon as i posted that last comment I went and read Entropy's comments on the msm8960 commit.  I abandoned it as you both make a lot of sense here.  I'm not sure if the failures are Verizon's doing or not, but the guy testing for jf on Verizon said it has failed every time.  I'll see if i can't find somewhere around campus tomorrow to test it out on ATT, but this *could* be a carrier thing as well.

Patch-set: 6
Label: Code-Review=0